fix flagd deps after mono repo split
