+++ providers/flagd/pkg/provider.go
@@ -7,8 +7,8 @@ import (
 	"fmt"
 	"github.com/go-logr/logr"
 	lru "github.com/hashicorp/golang-lru/v2"
-	flagdModels "github.com/open-feature/flagd/pkg/model"
-	flagdService "github.com/open-feature/flagd/pkg/service"
+	flagdModels "github.com/open-feature/flagd/core/pkg/model"
+	flagdService "github.com/open-feature/flagd/core/pkg/service"
 	"github.com/open-feature/go-sdk-contrib/providers/flagd/internal/logger"
 	"github.com/open-feature/go-sdk-contrib/providers/flagd/pkg/cache"
 	"github.com/open-feature/go-sdk-contrib/providers/flagd/pkg/constant"
