+++ providers/flagd/pkg/provider_cache_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	flagdModels "github.com/open-feature/flagd/pkg/model"
-	flagdService "github.com/open-feature/flagd/pkg/service"
+	flagdModels "github.com/open-feature/flagd/core/pkg/model"
+	flagdService "github.com/open-feature/flagd/core/pkg/service"
 	"github.com/open-feature/go-sdk-contrib/providers/flagd/internal/logger"
 	"github.com/open-feature/go-sdk-contrib/providers/flagd/internal/mock"
 	"github.com/open-feature/go-sdk-contrib/providers/flagd/pkg/constant"
