+++ tests/flagd/pkg/integration/caching.go
@@ -6,8 +6,8 @@ import (
 	"errors"
 	"fmt"
 	"github.com/cucumber/godog"
-	"github.com/open-feature/flagd/pkg/eval"
-	"github.com/open-feature/flagd/pkg/model"
+	"github.com/open-feature/flagd/core/pkg/eval"
+	"github.com/open-feature/flagd/core/pkg/model"
 	flagd "github.com/open-feature/go-sdk-contrib/providers/flagd/pkg"
 	"github.com/open-feature/go-sdk/pkg/openfeature"
 	"os"
